requirements.txt
pycairo==1.20.0
PyGObject==3.40.1
python-xlib==0.29
six==1.16.0
